+++ src/kernel_dal_echo/kernel/src/lib.rs
@@ -6,7 +6,6 @@
 use tezos_smart_rollup::entrypoint;
 use tezos_smart_rollup_debug::debug_msg;
 use tezos_smart_rollup_host::dal_parameters::RollupDalParameters;
-use tezos_smart_rollup_host::debug::HostDebug;
 use tezos_smart_rollup_host::path::OwnedPath;
 use tezos_smart_rollup_host::reveal::HostReveal;
 use tezos_smart_rollup_host::storage::StorageV1;
@@ -19,7 +18,7 @@ fn process_slot<Host>(
     page_size: usize,
     slot_index: u8,
 ) where
-    Host: HostReveal + StorageV1 + HostDebug,
+    Host: HostReveal + StorageV1,
 {
     let mut buffer = vec![0u8; page_size * num_pages];
 
@@ -91,7 +90,7 @@ fn get_slot_indexes_from_env() -> Vec<u8> {
 #[entrypoint::main]
 pub fn entry<Host>(host: &mut Host)
 where
-    Host: StorageV1 + WasmHost + HostDebug + HostReveal,
+    Host: StorageV1 + WasmHost + HostReveal,
 {
     let parameters = host.reveal_dal_parameters();
     debug_msg!(host, "Running kernel with parameters: {:?}\n", parameters);