+++ src/kernel_sdk/debug/src/lib.rs
@@ -5,64 +5,42 @@
 //! Provides *debug log* which can be written to, but does not affect the host state.
 //!
 //! The result of writing to the debug log is *implementation specific* - it may, for
-//! example, be written to a log file, or to `stdout` etc.
+//! example, be written to a log file, or to `stderr` etc.
 #![no_std]
 #![deny(missing_docs)]
 #![deny(rustdoc::broken_intra_doc_links)]
 
+// Re-exported so that the wasm macro body can reference it via `$crate::__core`
+// without requiring callers to depend on `tezos-smart-rollup-core` directly.
+#[cfg(pvm_kind = "wasm")]
+#[doc(hidden)]
+pub use tezos_smart_rollup_core as __core;
+
 /// Write a formatted message to host debug log. Formats follow [`core::fmt`].
 ///
-/// You can use `debug_msg!` with any variable such that implements [`HostDebug`].
-/// This is supported by any type implementing [`Runtime`] - such as
-/// [`RollupHost`].
+/// On WASM the message is written via the host's `write_debug` import. On all
+/// other targets it is written to `stderr` via `eprint!`.
 ///
-/// Using `debug_msg!` requires an allocator - so either you will need to include
-/// `extern crate alloc` when writing a kernel in a `no_std` context, or by
-/// depending on `std`. If you want to write debug logs without pulling in the
-/// allocator, use [debug_str] instead.
+/// The `host` argument is accepted for backward compatibility but is not used.
 ///
+/// # Example
 /// ```no_run
 /// extern crate alloc;
 /// use tezos_smart_rollup_debug::debug_msg;
-/// use tezos_smart_rollup_host::debug::HostDebug;
 ///
-/// fn log_runtime(host: &impl HostDebug) {
-///   debug_msg!(host, "Simple constant string");
-///
-///   debug_msg!(host, "A format {} with argument {}", "test", 5);
+/// fn log_something() {
+///   debug_msg!((), "Simple constant string");
+///   debug_msg!((), "A format {} with argument {}", "test", 5);
 /// }
 /// ```
-///
-/// # Compatability
-///
-/// For compatability with kernels that currently rely heavily on `Runtime` - the
-/// macros exposed here continue to correctly.
-///
-/// ```no_run
-/// extern crate alloc;
-/// use tezos_smart_rollup_debug::debug_msg;
-/// use tezos_smart_rollup_host::runtime::Runtime;
-///
-/// fn log_runtime(host: &impl Runtime) {
-///   debug_msg!(host, "Simple constant string");
-///
-///   debug_msg!(host, "A format {} with argument {}", "test", 5);
-/// }
-/// ```
-///
-/// [`HostDebug`]: tezos_smart_rollup_host::debug::HostDebug
-/// [`Runtime`]: tezos_smart_rollup_host::runtime::Runtime
-/// [`SmartRollupCore`]: tezos_smart_rollup_core::smart_rollup_core::SmartRollupCore
-/// [`RollupHost`]: tezos_smart_rollup_core::rollup_host::RollupHost
 #[cfg(feature = "alloc")]
 #[macro_export]
 macro_rules! debug_msg {
-    ($host: expr, $($args: expr),*) => {
-        {
-            extern crate alloc;
-            $crate::debug_str!($host, { &alloc::format!($($args), *) });
-        }
-    };
+    ($host: expr, $($args: expr),*) => {{
+        let _host = &$host;
+        extern crate alloc;
+        $crate::debug_str!($host, { &alloc::format!($($args),*) });
+    }};
 }
 
 /// Write a string to the debug log.
@@ -70,39 +48,54 @@ macro_rules! debug_msg {
 /// While this is less powerful than [`debug_msg`] (which supports format strings),
 /// it does _not_ require an allocator in order to be used.
 ///
-/// Just like `debug_msg`, you can use it with any variable implementing the [`HostDebug`]
-/// trait.
+/// On WASM the message is written via the host's `write_debug` import. On all
+/// other targets it is written to `stderr` via `eprint!`.
 ///
-/// use tezos_smart_rollup_debug::debug_str;
-/// use tezos_smart_rollup_host::debug::HostDebug;
+/// The `host` argument is accepted for backward compatibility but is not used.
 ///
-/// fn do_something(host: &impl HostDebug) {
-///   debug_str!(host, "Simple constant string");
-/// }
-///
-/// # Compatability
-///
-/// For compatability with kernels that currently rely heavily on [`Runtime`] - the
-/// macros exposed here continue to correctly.
-///
+/// # Example
 /// ```no_run
 /// use tezos_smart_rollup_debug::debug_str;
-/// use tezos_smart_rollup_host::runtime::Runtime;
 ///
-/// fn do_something(host: &impl Runtime) {
-///   debug_str!(host, "Simple constant string");
+/// fn do_something() {
+///   debug_str!((), "Simple constant string");
 /// }
 /// ```
-///
-/// [`Runtime`]: tezos_smart_rollup_host::runtime::Runtime
-/// [`HostDebug`]: tezos_smart_rollup_host::debug::HostDebug
+#[cfg(pvm_kind = "wasm")]
 #[macro_export]
 macro_rules! debug_str {
     ($host: expr, $msg: expr) => {{
-        use $crate::HostDebug;
-        $host.write_debug($msg);
+        let _host = &$host;
+        let msg: &str = $msg;
+        // SAFETY: msg is a valid UTF-8 string so its pointer/length are valid
+        unsafe { $crate::__core::target_impl::write_debug(msg.as_ptr(), msg.len()) };
     }};
 }
 
-#[doc(hidden)]
-pub use tezos_smart_rollup_host::debug::HostDebug;
+/// Write a string to the debug log.
+///
+/// While this is less powerful than [`debug_msg`] (which supports format strings),
+/// it does _not_ require an allocator in order to be used.
+///
+/// On WASM the message is written via the host's `write_debug` import. On all
+/// other targets it is written to `stderr` via `eprint!`.
+///
+/// The `host` argument is accepted for backward compatibility but is not used.
+///
+/// # Example
+/// ```no_run
+/// use tezos_smart_rollup_debug::debug_str;
+///
+/// fn do_something() {
+///   debug_str!((), "Simple constant string");
+/// }
+/// ```
+#[cfg(not(pvm_kind = "wasm"))]
+#[macro_export]
+macro_rules! debug_str {
+    ($host: expr, $msg: expr) => {{
+        let _host = &$host;
+        extern crate std;
+        std::eprint!("{}", $msg);
+    }};
+}